requirements.txt
streamlit
pandas
altair
matplotlib
seaborn
tqdm
python-dotenv